monkey patch maskrcnn single thread
