--- 3rd_party/Mask_RCNN_TF2/.gitignore
@@ -1,58 +1,0 @@
-# Data files and directories common in repo root
-datasets/
-logs/
-*.h5
-results/
-temp/
-test/
-
-# Byte-compiled / optimized / DLL files
-__pycache__/
-*.py[cod]
-*$py.class
-
-# Distribution / packaging
-.Python
-env/
-build/
-develop-eggs/
-dist/
-downloads/
-eggs/
-.eggs/
-lib/
-lib64/
-parts/
-sdist/
-var/
-wheels/
-*.egg-info/
-.installed.cfg
-*.egg
-
-# Installer logs
-pip-log.txt
-pip-delete-this-directory.txt
-
-# VS Studio Code
-.vscode
-
-# PyCharm
-.idea/
-
-# Dropbox
-.dropbox.attr
-
-# Jupyter Notebook
-.ipynb_checkpoints
-
-# pyenv
-.python-version
-
-# dotenv
-.env
-
-# virtualenv
-.venv
-venv/
-ENV/
