--- 3rd_party/Mask_RCNN_TF2/MANIFEST.in
@@ -1,3 +1,0 @@
-include README.md
-include LICENSE
-include requirements.txt
