--- 3rd_party/Mask_RCNN_TF2/requirements.txt
@@ -1,11 +1,0 @@
-numpy
-scipy
-Pillow
-cython
-matplotlib
-scikit-image
-tensorflow>=2.0.0
-opencv-python
-h5py
-imgaug
-IPython[all]
