--- 3rd_party/Mask_RCNN_TF2/setup.cfg
@@ -1,4 +1,0 @@
-[metadata]
-description-file = README.md
-license-file = LICENSE
-requirements-file = requirements.txt
